requirements.txt
Flask==2.2.3
redis==4.5.4
torch==1.12.1
transformers==4.21.1
pydub==0.25.1
flask-cors==3.0.10
python-dotenv==0.19.2